Remove tables package and move its functionality to dao
