--- src/main/scala/org/encryfoundation/explorer/db/tables/OutputsTable.scala
@@ -1,7 +1,0 @@
-package org.encryfoundation.explorer.db.tables
-
-object OutputsTable extends Table {
-  val name: String = "outputs"
-  val fields: Seq[String] = Seq("id", "tx_id", "monetary_value", "coin_id", "contract_hash", "data")
-  val fieldsString: String = fields.mkString("(", ", ", ")")
-}
