--- src/main/scala/org/encryfoundation/explorer/db/tables/Table.scala
@@ -1,7 +1,0 @@
-package org.encryfoundation.explorer.db.tables
-
-trait Table {
-  val name: String
-  val fields: Seq[String]
-  val fieldsString: String
-}
